updated queue

+++ pkg/queues/bq/bq.go
@@ -7,12 +7,14 @@ import (
 	"sync"
 	"time"
 
+	errs "github.com/ttn-nguyen42/taskq/internal/errors"
 	"github.com/ttn-nguyen42/taskq/internal/queue"
 	"go.etcd.io/bbolt"
 )
 
 const (
 	LeaseDuration = time.Minute
+	RetryTimeout  = time.Second * 30
 )
 
 type bqueue struct {
@@ -86,7 +88,7 @@ func (q *bqueue) Close() error {
 	return nil
 }
 
-func (q *bqueue) Ack(msgs queue.Messages) error {
+func (q *bqueue) Ack(queue string, id uint64) error {
 	q.mu.RLock()
 	bq := q.db
 	q.mu.RUnlock()
@@ -96,14 +98,12 @@ func (q *bqueue) Ack(msgs queue.Messages) error {
 	}
 
 	tx := func(tx *bbolt.Tx) error {
-		for _, msg := range msgs {
-			if err := q.ackSingle(tx, &msg); err != nil {
-				q.logger.
-					With("err", err).
-					With("met", "bqueue.Ack").
-					Error("failed to ack message")
-				return err
-			}
+		if err := q.ackSingle(tx, queue, id); err != nil {
+			q.logger.
+				With("err", err).
+				With("met", "bqueue.Ack").
+				Error("failed to ack message")
+			return err
 		}
 
 		return nil
@@ -116,12 +116,12 @@ func (q *bqueue) Ack(msgs queue.Messages) error {
 	return nil
 }
 
-func (q *bqueue) ackSingle(tx *bbolt.Tx, msg *queue.Message) error {
-	inProgressKey := queue.InProgressKey(msg.Queue)
-	msgKey := queue.MessageKey(msg.Queue, msg.ID)
+func (q *bqueue) ackSingle(tx *bbolt.Tx, name string, id uint64) error {
+	inProgressKey := queue.InProgressKey(name)
+	msgKey := queue.MessageKey(name, id)
 
 	statsKey := queue.StatsKey()
-	completedKey := queue.CompletedCountKey(msg.Queue)
+	completedKey := queue.CompletedCountKey(name)
 
 	inProgBucket, err := tx.CreateBucketIfNotExists(bytes(inProgressKey))
 	if err != nil {
@@ -138,17 +138,17 @@ func (q *bqueue) ackSingle(tx *bbolt.Tx, msg *queue.Message) error {
 		return fmt.Errorf("failed to create stats bucket: %w", err)
 	}
 
+	msg := inProgBucket.Get(bytes(msgKey))
+	if msg == nil {
+		return errs.NewErrAlreadyExists("message")
+	}
+
 	err = inProgBucket.Delete(bytes(msgKey))
 	if err != nil {
 		return fmt.Errorf("failed to delete message from in-progress: %w", err)
 	}
 
-	enc, err := queue.Encode(msg)
-	if err != nil {
-		return fmt.Errorf("failed to encode message: %w", err)
-	}
-
-	err = completedBucket.Put(bytes(msgKey), enc)
+	err = completedBucket.Put(bytes(msgKey), msg)
 	if err != nil {
 		return fmt.Errorf("failed to put message into completed: %w", err)
 	}
@@ -290,52 +290,56 @@ func (q *bqueue) dequeue(tx *bbolt.Tx, name string, opts *queue.DequeueOpts) (qu
 	return msgs, nil
 }
 
-func (q *bqueue) Enqueue(msgs queue.Messages) error {
+func (q *bqueue) Enqueue(msgs queue.Messages) (ids []uint64, err error) {
 	q.mu.RLock()
 	bq := q.db
 	q.mu.RUnlock()
 
 	if bq == nil {
-		return fmt.Errorf("queue is already shutdown")
+		return nil, fmt.Errorf("queue is already shutdown")
 	}
+
+	ids = make([]uint64, len(msgs))
 
 	tx := func(tx *bbolt.Tx) error {
 		for _, m := range msgs {
-			if err := q.enqueueSingle(tx, &m); err != nil {
+			id, err := q.enqueueSingle(tx, &m)
+			if err != nil {
 				q.logger.
 					With("err", err).
 					With("met", "bqeue.Enqueue").
 					Error("failed to enqueue message")
 				return err
 			}
+			ids = append(ids, id)
 		}
 		return nil
 	}
 
 	if err := bq.Update(tx); err != nil {
-		return fmt.Errorf("failed to update database messages: %w", err)
+		return nil, fmt.Errorf("failed to update database messages: %w", err)
 	}
 
-	return nil
+	return ids, nil
 }
 
-func (q *bqueue) enqueueSingle(tx *bbolt.Tx, msg *queue.Message) error {
+func (q *bqueue) enqueueSingle(tx *bbolt.Tx, msg *queue.Message) (id uint64, err error) {
 	enc, err := queue.Encode(msg)
 	if err != nil {
-		return fmt.Errorf("failed to encode message: %w", err)
+		return 0, fmt.Errorf("failed to encode message: %w", err)
 	}
 
 	pendingKey := queue.PendingKey(msg.Queue)
 
 	pending, err := tx.CreateBucketIfNotExists(bytes(pendingKey))
 	if err != nil {
-		return fmt.Errorf("failed to create bucket: %w", err)
+		return 0, fmt.Errorf("failed to create bucket: %w", err)
 	}
 
 	if msg.ID == 0 {
 		msgId, err := pending.NextSequence()
 		if err != nil {
-			return fmt.Errorf("failed to get next sequence: %w", err)
+			return 0, fmt.Errorf("failed to get next sequence: %w", err)
 		}
 		msg.ID = msgId
 	}
@@ -344,77 +348,92 @@ func (q *bqueue) enqueueSingle(tx *bbolt.Tx, msg *queue.Message) error {
 
 	err = pending.Put(bytes(msgKey), enc)
 	if err != nil {
-		return fmt.Errorf("failed to put message: %w", err)
+		return 0, fmt.Errorf("failed to put message: %w", err)
 	}
 
-	return nil
+	return msg.ID, nil
 }
 
-func (q *bqueue) Requeue(msgs queue.Messages) error {
+func (q *bqueue) Requeue(queue string, id uint64) (newId uint64, err error) {
 	q.mu.RLock()
 	bq := q.db
 	q.mu.RUnlock()
 
 	if bq == nil {
-		return fmt.Errorf("queue is already shutdown")
+		return 0, fmt.Errorf("queue is already shutdown")
 	}
 
 	tx := func(tx *bbolt.Tx) error {
-		for _, msg := range msgs {
-			if err := q.requeueSingle(tx, msg); err != nil {
-				q.logger.
-					With("err", err).
-					With("met", "bqueue.Requeue").
-					Error("failed to requeue message")
-				return err
-			}
+		newId, err = q.requeueSingle(tx, queue, id)
+		if err != nil {
+			q.logger.
+				With("err", err).
+				With("met", "bqueue.Requeue").
+				Error("failed to requeue message")
+			return err
 		}
 
 		return nil
 	}
 
 	if err := bq.Update(tx); err != nil {
-		return fmt.Errorf("failed to update database messages: %w", err)
+		return 0, fmt.Errorf("failed to update database messages: %w", err)
 	}
 
-	return nil
+	return 0, nil
 }
 
-func (q *bqueue) requeueSingle(tx *bbolt.Tx, msg queue.Message) error {
-	pending := queue.PendingKey(msg.Queue)
-	inProgress := queue.InProgressKey(msg.Queue)
-	msgKey := queue.MessageKey(msg.Queue, msg.ID)
+func (q *bqueue) requeueSingle(tx *bbolt.Tx, name string, id uint64) (newId uint64, err error) {
+	pending := queue.PendingKey(name)
+	inProgress := queue.InProgressKey(name)
+	msgKey := queue.MessageKey(name, id)
 
 	inProgBucket, err := tx.CreateBucketIfNotExists(bytes(inProgress))
 	if err != nil {
-		return fmt.Errorf("failed to create in-progress bucket: %w", err)
+		return 0, fmt.Errorf("failed to create in-progress bucket: %w", err)
 	}
 
 	pendingBucket, err := tx.CreateBucketIfNotExists(bytes(pending))
 	if err != nil {
-		return fmt.Errorf("failed to create pending bucket: %w", err)
+		return 0, fmt.Errorf("failed to create pending bucket: %w", err)
 	}
 
-	enc, err := queue.Encode(&msg)
+	msg := inProgBucket.Get(bytes(msgKey))
+	if msg == nil {
+		return 0, errs.NewErrNotFound("message")
+	}
+
+	dec, err := queue.Decode(msg)
 	if err != nil {
-		return fmt.Errorf("failed to encode message: %w", err)
+		return 0, fmt.Errorf("failed to decode message: %w", err)
+	}
+
+	newId, err = pendingBucket.NextSequence()
+	if err != nil {
+		return 0, fmt.Errorf("failed to get next sequence: %w", err)
+	}
+	dec.ID = newId
+
+	enc, err := queue.Encode(dec)
+	if err != nil {
+		return 0, fmt.Errorf("failed to encode message: %w", err)
 	}
 
 	err = inProgBucket.Delete(bytes(msgKey))
 	if err != nil {
-		return fmt.Errorf("failed to delete message from in-progress: %w", err)
+		return 0, fmt.Errorf("failed to delete message from in-progress: %w", err)
 	}
 
 	pendingBucket.Put(bytes(msgKey), enc)
 
-	return nil
+	return newId, nil
 }
 
-func (q *bqueue) Discard(msg queue.Messages) error {
+func (q *bqueue) Discard(queue string, id uint64) error {
 	panic("unimplemented")
 }
 
-func (q *bqueue) Retry(msgs queue.Messages) error {
+func (q *bqueue) Retry(queue string, id uint64) error {
 	q.mu.RLock()
 	bq := q.db
 	q.mu.RUnlock()
@@ -424,14 +443,12 @@ func (q *bqueue) Retry(msgs queue.Messages) error {
 	}
 
 	tx := func(tx *bbolt.Tx) error {
-		for _, msg := range msgs {
-			if err := q.retrySingle(tx, msg); err != nil {
-				q.logger.
-					With("err", err).
-					With("met", "bqueue.Retry").
-					Error("failed to retry message")
-				return err
-			}
+		if err := q.retrySingle(tx, queue, id); err != nil {
+			q.logger.
+				With("err", err).
+				With("met", "bqueue.Retry").
+				Error("failed to retry message")
+			return err
 		}
 
 		return nil
@@ -444,10 +461,10 @@ func (q *bqueue) Retry(msgs queue.Messages) error {
 	return nil
 }
 
-func (q *bqueue) retrySingle(tx *bbolt.Tx, msg queue.Message) error {
-	inProgressKey := queue.InProgressKey(msg.Queue)
-	retryKey := queue.RetryKey(msg.Queue)
-	msgKey := queue.MessageKey(msg.Queue, msg.ID)
+func (q *bqueue) retrySingle(tx *bbolt.Tx, name string, id uint64) error {
+	inProgressKey := queue.InProgressKey(name)
+	retryKey := queue.RetryKey(name)
+	msgKey := queue.MessageKey(name, id)
 
 	inProgBucket, err := tx.CreateBucketIfNotExists(bytes(inProgressKey))
 	if err != nil {
@@ -463,7 +480,20 @@ func (q *bqueue) retrySingle(tx *bbolt.Tx, msg queue.Message) error {
 		return fmt.Errorf("failed to delete message from in-progress: %w", err)
 	}
 
-	enc, err := queue.Encode(&msg)
+	msg := inProgBucket.Get(bytes(msgKey))
+	if msg == nil {
+		return errs.NewErrNotFound("message")
+	}
+
+	dec, err := queue.Decode(msg)
+	if err != nil {
+		return fmt.Errorf("failed to decode message: %w", err)
+	}
+
+	dec.Retried += 1
+	dec.LastFailedAt = time.Now()
+
+	enc, err := queue.Encode(dec)
 	if err != nil {
 		return fmt.Errorf("failed to encode message: %w", err)
 	}
@@ -609,6 +639,172 @@ func (q *bqueue) getPendingCount(tx *bbolt.Tx, name string) (uint64, error) {
 	}
 
 	return uint64(pending.Stats().KeyN), nil
+}
+
+func (q *bqueue) Move(id uint64, from string, to string) (newId uint64, err error) {
+	q.mu.RLock()
+	bq := q.db
+	q.mu.RUnlock()
+
+	if bq == nil {
+		return 0, fmt.Errorf("queue is already shutdown")
+	}
+
+	tx := func(tx *bbolt.Tx) error {
+		newId, err = q.moveSingle(tx, id, from, to)
+		if err != nil {
+			q.logger.
+				With("err", err).
+				With("met", "bqueue.Move").
+				Error("failed to move message")
+			return err
+		}
+
+		return nil
+	}
+
+	if err := bq.Update(tx); err != nil {
+		return 0, fmt.Errorf("failed to update database messages: %w", err)
+	}
+
+	return newId, nil
+}
+
+func (q *bqueue) moveSingle(tx *bbolt.Tx, id uint64, from, to string) (newId uint64, err error) {
+	fromKey := queue.PendingKey(from)
+	toKey := queue.PendingKey(to)
+	msgKey := queue.MessageKey(from, id)
+
+	fromBucket := tx.Bucket(bytes(fromKey))
+	if fromBucket == nil {
+		return 0, fmt.Errorf("source queue not found")
+	}
+
+	toBucket := tx.Bucket(bytes(toKey))
+	if toBucket == nil {
+		return 0, fmt.Errorf("destination queue not found")
+	}
+
+	raw := fromBucket.Get(bytes(msgKey))
+	if raw == nil {
+		return 0, fmt.Errorf("message not found")
+	}
+
+	newId, err = toBucket.NextSequence()
+	if err != nil {
+		return 0, fmt.Errorf("failed to get next sequence: %w", err)
+	}
+
+	newMsgKey := queue.MessageKey(to, newId)
+
+	err = toBucket.Put(bytes(newMsgKey), raw)
+	if err != nil {
+		return 0, fmt.Errorf("failed to put message into destination queue: %w", err)
+	}
+
+	err = fromBucket.Delete(bytes(msgKey))
+	if err != nil {
+		return 0, fmt.Errorf("failed to delete message from source queue: %w", err)
+	}
+
+	return newId, nil
+}
+
+func (q *bqueue) ReconcileRetry(limit int, queues ...string) (ids []uint64, newIds []uint64, err error) {
+	q.mu.RLock()
+	bq := q.db
+	q.mu.RUnlock()
+
+	if bq == nil {
+		return nil, nil, fmt.Errorf("queue is already shutdown")
+	}
+
+	tx := func(tx *bbolt.Tx) error {
+		for _, qu := range queues {
+			ids, newIds, err = q.reconcileRetryQueue(tx, qu, limit)
+			if err != nil {
+				q.logger.
+					With("err", err).
+					With("met", "bqueue.reconcileRetry").
+					Error("failed to reconcile retry queue")
+				return err
+			}
+		}
+
+		return nil
+	}
+
+	if err := bq.Update(tx); err != nil {
+		return nil, nil, fmt.Errorf("failed to update database messages: %w", err)
+	}
+
+	return ids, newIds, nil
+}
+
+func (q *bqueue) reconcileRetryQueue(tx *bbolt.Tx, name string, limit int) (ids []uint64, newIds []uint64, err error) {
+	retryKey := queue.RetryKey(name)
+	pendingKey := queue.PendingKey(name)
+
+	retryBucket, err := tx.CreateBucketIfNotExists(bytes(retryKey))
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to create retry bucket: %w", err)
+	}
+
+	pendingBucket, err := tx.CreateBucketIfNotExists(bytes(pendingKey))
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to create pending bucket: %w", err)
+	}
+
+	retryCur := retryBucket.Cursor()
+
+	itemsReconciled := 0
+	ids = make([]uint64, 0, limit)
+	newIds = make([]uint64, 0, limit)
+
+	for key, val := retryCur.First(); key != nil; key, val = retryCur.Next() {
+		msg, err := queue.Decode(val)
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to decode message: %w", err)
+		}
+
+		err = retryBucket.Delete(key)
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to delete message from retry: %w", err)
+		}
+
+		if msg.Retried >= msg.MaxRetry {
+			continue
+		}
+
+		ids = append(ids, msg.ID)
+
+		newId, err := pendingBucket.NextSequence()
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to get next sequence: %w", err)
+		}
+		msg.ID = newId
+
+		newIds = append(newIds, newId)
+
+		newKey := queue.MessageKey(name, newId)
+
+		enc, err := queue.Encode(msg)
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to encode message: %w", err)
+		}
+
+		err = pendingBucket.Put(bytes(newKey), enc)
+		if err != nil {
+			return nil, nil, fmt.Errorf("failed to put message into pending: %w", err)
+		}
+
+		itemsReconciled += 1
+		if itemsReconciled >= limit {
+			break
+		}
+	}
+
+	return ids, newIds, nil
 }
 
 func bytes(s string) []byte {
